Fix types issue
