+++ packages/db/package.json
@@ -1,16 +1,20 @@
 {
   "name": "@repo/db",
-  "module": "./src/index.ts",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
+    "build": "tsc",
     "drizzle:generate": "drizzle-kit generate",
     "drizzle:push": "drizzle-kit push",
     "drizzle:studio": "drizzle-kit studio",
     "drizzle:migrate": "drizzle-kit migrate"
   },
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./dist/index.js"
   },
-  "types": "./src/index.ts",
+  "files": [
+    "dist"
+  ],
   "type": "module",
   "devDependencies": {
     "@types/bun": "latest",
