+++ packages/db/src/index.ts
@@ -12,3 +12,4 @@ export { slices, sliceRelations } from "./schema/slice";
 export { sliceAccess, sliceAccessRelations } from "./schema/sliceAccess";
 export { memos, memoRelations, memoTypeEnum } from "./schema/memo";
 export { refreshTokens, refreshTokenRelations } from "./schema/refreshToken";
+export * from "./types";
