+++ packages/types/src/common/auth.ts
