+++ packages/types/src/common/validation/auth.schema.ts
