+++ packages/types/src/common/validation/index.ts
@@ -1,3 +1,3 @@
 export * from "./auth.schema";
-export * from "./slice.schema";
 export * from "./memo.schema";
+export * from "./slice.schema";
