+++ packages/types/src/common/validation/memo.schema.ts
