+++ packages/types/src/common/validation/slice.schema.ts
